cc: Use fastcall bpf_rdonly_cast()
Since torvalds/linux@da7d71b ("bpf: Use KF_FASTCALL to mark kfuncs supporting fastcall contract"),
`bpf_rdonly_cast()` kfunc has been marked with flag `KF_FASTCALL`, which
means the unused registers R3, R4 and R5 won't be volatile after calling
`bpf_rdonly_cast()` kfunc.
